changes after Wojteks comments 1
